move alert type enum to a separate file
