+++ src/app/components/alert/alert-type.enum.ts
@@ -1,0 +1,7 @@
+export enum AlertType
+{
+    info = 'info',
+    warning = 'warning',
+    error = 'danger',
+    success = 'success'
+}
